feat: Do not exit from default Python script
When a new App is created, it is populated with a simple Python script and "bare minimum" Arduino sketch.

For initial explorations of a new system, or when troubleshooting, it is common to create very simple programs (AKA
"Hello, world!"). Although a complex App will typically consist of a Python script and Arduino sketch program working in
coordination, for users with prior experience with Arduino the natural approach to creating a minimal App will be to
simply write some familiar Arduino sketch code, leaving the default Python script code as-is.

The App is considered to be in a "stopped" state as soon as the Python script exits. This is the correct approach, but
may be confusing to users due to the fact that, except perhaps under exceptional conditions, the Arduino sketch program
runs perpetually. The author of a minimal sketch-based "Hello, world!" App will find it unintuitive if their App goes
into a "stopped" state immediately after starting. The previous default Python script produced exactly that result.

The default Python script is hereby changed to the more intuitive behavior of running perpetually.

Co-authored-by: Davide <davideneri18@gmail.com>